AIPOC/requirements.txt
streamlit
pydantic>=2
pypdf
rapidfuzz
regex
requests>=2.28.0
pdfplumber
pytesseract
langchain-core
langchain-text-splitters
langchain-anthropic
langchain-chroma
langchain-community
sentence-transformers
chromadb
python-docx
pytest
reportlab



